refactor: add index.js for each graphql folder
