+++ src/graphql/subscriptions/index.js
@@ -1,0 +1,10 @@
+/**
+ * @file Contains all GraphQL subscription exports.
+ * @author Manuel Cabral
+ * @version 0.0.1
+ */
+
+module.exports = {
+	chatMessageAdded: require('./chatMessageAdded'),
+	messageAdded: require('./messageAdded'),
+}
