--- src/graphql/types.js
@@ -1,11 +1,0 @@
-/**
- * @file Contains all GraphQL object types.
- * @author Manuel Cabral
- * @version 0.0.6
- */
-
-module.exports = {
-	UserType: require('./types/userType'),
-	MessageType: require('./types/messageType'),
-	ChatType: require('./types/chatType'),
-}
