+++ src/graphql/types/index.js
@@ -1,0 +1,11 @@
+/**
+ * @file Contains all GraphQL object types.
+ * @author Manuel Cabral
+ * @version 0.0.6
+ */
+
+module.exports = {
+	UserType: require('./userType'),
+	MessageType: require('./messageType'),
+	ChatType: require('./chatType'),
+}
